updating user

+++ mobile/screens/profile/profileInfo.tsx
@@ -1,50 +1,68 @@
 import { yupResolver } from "@hookform/resolvers/yup";
-import { useContext, useState } from "react";
+import { useContext } from "react";
 import { Controller, useForm } from "react-hook-form";
-import { TouchableOpacity } from "react-native";
-import { Button, Div, Icon, Input, Text } from "react-native-magnus";
+import { Button, Div, Input, Text } from "react-native-magnus";
 import { verticalScale } from "react-native-size-matters";
 import * as yup from "yup";
 import { ErrorInputMessageContainer, ErrorMessageText, LabelContainer } from "../../components/styled/styled";
 import { AuthContext } from "../../context/authProvider";
+import { useGlobalUI } from "../../context/GlobalUIProvider";
+import { useMutate } from "../../hooks/useMutate";
 import authService from "../../service/auth.service";
 import userService from "../../service/user.service";
-import { UserInfoRegister } from "../../types/user.type";
+import { UpdateUserInfo } from "../../types/user.type";
 
 const validationSchema = yup.object({
   name: yup.string().required("Requerido").min(8, 'El nombre de usuario debe tener al menos 8 caracteres'),
   email: yup.string().required("Requerido").email('Debe ser un email válido'),
-  password: yup.string().required("Requerido").min(8, 'La contraseña debe tener al menos 8 caracteres'),
 });
 
-const ProfileInfo = () => {
-  const [visibility, setVisibility] = useState(true);
+interface userInfo {
+  email: string,
+  name: string
+}
 
-  const toggleVisibility = () => {
-    setVisibility(!visibility);
-  };
+const ProfileInfo = () => {
 
   const { setCurrentUser, currentUser } = useContext(AuthContext);
+  const { showSnackBar } = useGlobalUI()
 
   async function handleLogout() {
     await authService.signOut();
     setCurrentUser(null);
   }
+
   async function handleDelete() {
+    if (!currentUser) return
     await userService.deleteUser(currentUser.id);
     setCurrentUser(null);
+  }
+
+  async function handleUpdate(data: UpdateUserInfo) {
+    if (!currentUser) return
+    await userService.updateUser({ id: currentUser.id, email: data.email && data.email, name: data.name && data.name })
+  }
+
+  const onError = (e) => {
+    showSnackBar('error', 'Hubo un problema')
+    console.log(e)
+  }
+
+  const updateQuery = useMutate(handleUpdate, () => showSnackBar('success', 'Guardado con exito'), false, (e) => onError(e))
+
+  const onSubmit = async (data: UpdateUserInfo) => {
+    await updateQuery(data)
   }
 
   const {
     control,
     handleSubmit,
     formState: { errors },
-  } = useForm<UserInfoRegister>({
+  } = useForm<userInfo>({
     resolver: yupResolver(validationSchema),
     defaultValues: {
       name: currentUser?.name || '',
-      email: currentUser?.email || '',
-      password: ''
+      email: currentUser?.email || ''
     }
   })
 
@@ -105,43 +123,13 @@ const ProfileInfo = () => {
           )}
           name="email"
         />
-        <Controller
-          control={control}
-          render={({ field: { onChange, value } }) => (
-            <>
-              <LabelContainer alignSelf="flex-start" mb='xs'>
-                <Text color='secondary'>Contraseña</Text>
-              </LabelContainer>
-              <Input
-                fontSize={'sm'}
-                placeholder="**********"
-                placeholderTextColor={"black"}
-                onChangeText={onChange}
-                value={value}
-                h={verticalScale(35)}
-                mb='lg'
-                secureTextEntry={visibility}
-                suffix={
-                  <TouchableOpacity onPress={toggleVisibility}>
-                    <Icon name={visibility ? "eye" : 'eye-off'} color="gray900" fontFamily="Feather" />
-                  </TouchableOpacity>
-                }
-              />
-              <ErrorInputMessageContainer>
-                {errors.password && <ErrorMessageText>{errors.password?.message as string}</ErrorMessageText>}
-              </ErrorInputMessageContainer>
-            </>
-          )}
-          name="password"
-        />
       </Div>
       <Div>
-        <Button bg="secondary" color="black" w={"100%"}>
+        <Button bg="secondary" color="black" w={"100%"} onPress={handleSubmit(onSubmit)}>
           GUARDAR
         </Button>
         <Div my={"xl"} flexDir="row" justifyContent="space-between">
           <Text onPress={handleLogout}>Cerrar sesion</Text>
-
           <Text onPress={handleDelete}>Eliminar cuenta</Text>
         </Div>
       </Div>
